Move testing objects to the tests root
